import: routed.kicad_* (manually routed PCB) [import]
Commit the .kicad_pcb and related files after manually adding the
copper traces to the PCB in KiCad (i.e. the "routing" step).

Note 1: Ergogen does not perform the routing step because it's too
hard to automate. Therefore I had to do the routing manually in
KiCad. That also means that if I ever make further changes to
`ergogen.yaml` and regenerate, I will need to redo the routing!

Note 2: This is my first time doing routing in KiCad and it took about
2 hours. Next time, I expect I could complete it under an hour.